pyproject.toml
[project]
name = "lastfmeventscrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "pyyaml>=6.0.3",
    "python-dotenv>=1.2.1",
    "jinja2>=3.1.6",
]
